+++ src/cogs/general/storage/model/guild_member.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, ForeignKey, BigInteger
 from sqlalchemy.orm import relationship
 
 from src.database import Base
@@ -9,8 +9,8 @@ class GuildMember(Base):
 
     id = Column(Integer, primary_key=True)
 
-    guild_id = Column(Integer, ForeignKey("guild.id"))
-    member_id = Column(Integer, ForeignKey("member.id"))
+    guild_id = Column(BigInteger, ForeignKey("guild.id"))
+    member_id = Column(BigInteger, ForeignKey("member.id"))
 
     guild = relationship("Guild", back_populates="members")
     member = relationship("Member", back_populates="guilds")
